Refactored Truffle Tests (#47)
* significant refactoring

* updated successfulWithdraw utility

* moved repeating code to utilities.js

* file rename, util var for all functions

* fixed last test bug

* separated utilities and rootchain helpers

* removed testing code from rootchain contract

* removed test code and destructured utilities

* formatting fix
